requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.12
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0